pyproject.toml
[tool.poetry]
name = "discord-ai-assistant-bot"
version = "0.1.0"
description = "Discord bot using Cogs with Google Gemini integration (/q, /imginfo, /fix_grammar, /voice)"
authors = ["Jialong0209 <smt02091395@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
"discord.py" = "^2.4.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"
edge-tts = "^6.1.12"
pynacl = "^1.5.0"
pandas = "^2.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.5.6"
mypy = "^1.10.0"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

